combine import config tests
